Build NPM using DNT And consolidate the build & release jobs using a matrix.
