+++ .github/workflows/test-and-release.yml
@@ -40,37 +40,6 @@ jobs:
           fail_ci_if_error: true
           verbose: true
 
-  build:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Git Checkout
-        uses: actions/checkout@v2
-
-      - uses: actions/setup-node@v2
-        with:
-          node-version: 16
-          registry-url: 'https://npm.pkg.github.com'
-
-      - name: Setup Deno
-        uses: denoland/setup-deno@v1
-        with:
-          deno-version: v1.x
-
-      - name: Install
-        run: make install
-
-      - name: Build
-        run: make all
-
-      - name: Post Build Test
-        run: make post-build-test
-
-      - uses: actions/upload-artifact@v2
-        with:
-          name: dist
-          path: dist
-
   version:
     runs-on: ubuntu-latest
     if: startsWith(github.ref, 'refs/tags/v')
@@ -85,44 +54,62 @@ jobs:
       version: ${{ steps.get_version.outputs.version }}
       is_pre: ${{ steps.is_pre.outputs.IS_PRE }}
 
-  package-json:
-    runs-on: ubuntu-latest
+  build:
     needs: [version]
+    strategy:
+      matrix:
+        registry: [npm, github]
+        include:
+          - registry: npm
+            scope: "eyalsh"
+            registry-url: https://registry.npmjs.org
+          - registry: github
+            scope: "Eyal-Shalev"
+            registry-url: 'https://npm.pkg.github.com'
+    runs-on: ubuntu-latest
+
     steps:
       - name: Git Checkout
         uses: actions/checkout@v2
+
+      - uses: actions/setup-node@v2
+        with:
+          node-version: 16
+          registry-url: ${{ matrix.registry-url }}
 
       - name: Setup Deno
         uses: denoland/setup-deno@v1
         with:
           deno-version: v1.x
 
-      - name: Make GitHub package.json
-        env:
-          SCOPE: Eyal-Shalev
+      - name: Install
+        run: make install
+
+      - name: Build
+        run: make build
+        with:
           VERSION: ${{ needs.version.outputs.version }}
-        run: make package-json
+          SCOPE: ${{ matrix.scope }}
 
       - uses: actions/upload-artifact@v2
         with:
-          name: github-package-json
-          path: package.json
+          name: dist-${{ matrix.registry }}
+          path: dist
 
-      - name: Make NPM package.json
-        env:
-          SCOPE: eyalsh
-          VERSION: ${{ needs.version.outputs.version }}
-        run: make package-json
-
-      - uses: actions/upload-artifact@v2
-        with:
-          name: npm-package-json
-          path: package.json
-
-  gpr-publish:
-    name: GitHub Publish
+  publish:
+    name: Publish
+    needs: [test, build]
+    strategy:
+      matrix:
+        registry: [npm, github]
+        include:
+          - registry: npm
+            registry-url: https://registry.npmjs.org
+            token: ${{ secrets.NPM_TOKEN }}
+          - registry: github
+            registry-url: 'https://npm.pkg.github.com'
+            token: ${{ secrets.GITHUB_TOKEN }}
     runs-on: ubuntu-latest
-    needs: [test, build, package-json]
     permissions:
       packages: write
       contents: read
@@ -133,53 +120,18 @@ jobs:
 
       - uses: actions/download-artifact@v2
         with:
-          name: dist
+          name: dist-${{ matrix.registry }}
           path: dist
-
-      - uses: actions/download-artifact@v2
-        with:
-          name: github-package-json
-          path: "."
 
       - uses: actions/setup-node@v2
         with:
           node-version: 16
-          registry-url: 'https://npm.pkg.github.com'
+          registry-url: ${{ matrix.registry-url }}
 
       - run: npm publish
+        working-directory: dist/npm
         env:
-          NODE_AUTH_TOKEN: ${{secrets.GITHUB_TOKEN}}
-
-  npm-publish:
-    name: NPM Publish
-    runs-on: ubuntu-latest
-    needs: [test, build, package-json]
-    permissions:
-      packages: write
-      contents: read
-
-    steps:
-      - name: Git Checkout
-        uses: actions/checkout@v2
-
-      - uses: actions/download-artifact@v2
-        with:
-          name: dist
-          path: dist
-
-      - uses: actions/download-artifact@v2
-        with:
-          name: npm-package-json
-          path: "."
-
-      - uses: actions/setup-node@v2
-        with:
-          node-version: 16
-          registry-url: 'https://registry.npmjs.org'
-
-      - run: npm publish
-        env:
-          NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
+          NODE_AUTH_TOKEN: ${{ matrix.token }}
 
   release:
     name: Release
@@ -205,8 +157,8 @@ jobs:
           files: |
             LICENSE
             README.md
-            dist/*.js
-            dist/*.js
+            dist/iife.js
+            dist/iife.min.js
 
       - uses: marvinpinto/action-automatic-releases@latest
         if: ${{ ! needs.version.outputs.is_pre }}
@@ -217,5 +169,5 @@ jobs:
           files: |
             LICENSE
             README.md
-            dist/*.js
-            dist/*.js
+            dist/iife.js
+            dist/iife.min.js
