+++ .gitignore
@@ -5,5 +5,4 @@
 /coverage
 /node_modules
 /package-lock.json
-/benchmark.txt
-/benchmark.html
+/benchmark.*
