+++ src/time_test.ts
@@ -69,10 +69,11 @@ Deno.test("after", async () => {
 });
 
 async function analyzeTicker(interval: number, times: number) {
-  const startMs = new Date().getTime();
+  const data = Array(times);
   const ticker = new Ticker(interval);
   const results: number[] = [];
-  for (const _ of Array(times)) {
+  const startMs = new Date().getTime();
+  for (const _ of data) {
     const [val, _] = await ticker.c.get();
     assert(val);
     results.push(val.getTime());
@@ -81,18 +82,17 @@ async function analyzeTicker(interval: number, times: number) {
   const resultsWithStart = [startMs, ...results];
   const intervals = results.map((next, index) =>
     (next - resultsWithStart[index]) - interval
-  );
+  ).sort();
 
+  const sum = intervals.reduce((acc, item) => acc + item);
   return {
-    avg: intervals.reduce((acc, item) => acc + item) / times,
-    min: Math.min(...intervals),
-    max: Math.max(...intervals),
+    p95: intervals[Math.floor(intervals.length * 0.95)],
+    avg: sum / times,
   };
 }
 
 Deno.test("Ticker", async () => {
-  const { avg, min, max } = await analyzeTicker(30, 50);
+  const { avg, p95 } = await analyzeTicker(10, 100);
   assertLessThan(avg, 2);
-  assertLessThan(min, 2);
-  assertLessThan(max, 3.01);
+  assertLessThan(p95, 2.01);
 });
