+++ src/time_tick_test.ts
@@ -1,10 +1,12 @@
 import { tick } from "./time.ts";
 import { assertNumberBetween } from "./internal/test_utils.ts";
+import { isDeno } from "which_runtime";
 
 // Note: This test was moved outside of time_test.ts because it leaks ops and causes other tests to fail.
 
 Deno.test({
   name: "tick",
+  ignore: !isDeno,
   // This test is expected to leak ops & resources because there is no way to stop `tick`.
   sanitizeOps: false,
   sanitizeResources: false,
